Inject image-map at runtime
This is the second part of the move to the image map injection move.
Now that we have a map from logical-name to an image definition
(containing the image reference, digest, description etc.), we inject it
in the invocation image in the same way we deal with parameters and
credentials.

The invocation image will then be able to consume this map to replace
any utility image reference declared in the bundle.

Signed-off-by: Simon Ferquel <simon.ferquel@docker.com>